requirements.txt
Flask==2.3.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Werkzeug==2.3.0
pymysql==1.1.0
cryptography==41.0.7
bcrypt==4.1.2
pydantic==2.10.4
tzlocal==5.2
pytz==2024.2